requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.5
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
cron-descriptor==1.4.0
cryptography==41.0.5
Django==4.2.7
django-celery-beat==2.5.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
factory-boy==3.3.0
Faker==20.0.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.3
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.40
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.3
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
